Perkeltas duomenų fetchinimas į atskirą klasę iš kontrolerio prisijungus paieškos
